+++ src/scoreboard.h
@@ -10,8 +10,8 @@
 ///
 ///
 typedef struct Scoreboard {
-    u8 player_1_score;
-    u8 player_2_score;
+    usize player_1_score;
+    usize player_2_score;
     const char *player_1_name;
     const char *player_2_name;
 } Scoreboard;
@@ -24,11 +24,11 @@ void SB_redraw(Scoreboard *sb, const float ui_padding, Color background_color);
 ///
 ///
 ///
-void update_player_1_score(Scoreboard *sb, u8 score);
+void update_player_1_score(Scoreboard *sb, usize score);
 
 ///
 ///
 ///
-void update_player_2_score(Scoreboard *sb, u8 score);
+void update_player_2_score(Scoreboard *sb, usize score);
 
 #endif
